+++ Locales/Locale-enUS.lua
@@ -80,6 +80,7 @@ L["Description"] = true
 L["Import"] = true
 L["Tex Coords"] = true
 L["Use Skin"] = true
+L["Materials"] = true
 
 -- Modules\ObjectiveBuilder.lua
 L.FilterAutoItemsHint = [[Check this option to hide automatically created item objectives (prepended by "item:").]]